requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
azure-ai-ml==1.26.2
azure-common==1.1.28
azure-core==1.33.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.21.0
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.7
azure-monitor-opentelemetry-exporter==1.0.0b36
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
backoff==2.2.1
beautifulsoup4==4.13.4
browser-use==0.1.40
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
defusedxml==0.7.1
deprecated==1.2.18
distro==1.9.0
fastapi==0.115.12
fixedint==0.1.6
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib-metadata==8.6.1
isodate==0.7.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain-anthropic==0.3.3
langchain-core==0.3.52
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.31
markdownify==0.14.1
marshmallow==4.0.0
monotonic==1.6
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.2.2
ollama==0.4.8
openai==1.75.0
opentelemetry-api==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-dbapi==0.53b1
opentelemetry-instrumentation-django==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-instrumentation-flask==0.53b1
opentelemetry-instrumentation-psycopg2==0.53b1
opentelemetry-instrumentation-requests==0.53b1
opentelemetry-instrumentation-urllib==0.53b1
opentelemetry-instrumentation-urllib3==0.53b1
opentelemetry-instrumentation-wsgi==0.53b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
packaging==24.2
playwright==1.51.0
posthog==3.25.0
psutil==6.1.1
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydash==8.0.5
pyee==12.1.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.24.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
strictyaml==1.7.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
pandas==2.2.3
